requirements.txt
anyio==4.10.0
certifi==2025.8.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
psycopg2-binary==2.9.10
python-dotenv==1.1.1
python-telegram-bot==21.5
sniffio==1.3.1
typing_extensions==4.15.0
flask
flask
psycopg2-binary
python-dotenv
python-telegram-bot
telegram